move a custom hook
